refactor: JpaRepository 패키지 분리
