+++ src/main/java/kitchenpos/menu/repository/JpaMenuGroupRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu.repository;
 
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroupRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
