+++ src/main/java/kitchenpos/menu/repository/JpaMenuRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu.repository;
 
+import kitchenpos.menu.domain.Menu;
+import kitchenpos.menu.domain.MenuRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
