+++ src/main/java/kitchenpos/order/repository/JpaOrderRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.repository;
 
+import kitchenpos.order.domain.Order;
+import kitchenpos.order.domain.OrderRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
