+++ src/main/java/kitchenpos/order/repository/JpaOrderTableRepository.java
@@ -1,5 +1,7 @@
-package kitchenpos.order.domain;
+package kitchenpos.order.repository;
 
+import kitchenpos.order.domain.OrderTable;
+import kitchenpos.order.domain.OrderTableRepository;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
