+++ src/test/java/kitchenpos/menu/repository/InMemoryMenuGroupRepository.java
@@ -1,4 +1,7 @@
-package kitchenpos.menu.domain;
+package kitchenpos.menu.repository;
+
+import kitchenpos.menu.domain.MenuGroup;
+import kitchenpos.menu.domain.MenuGroupRepository;
 
 import java.util.ArrayList;
 import java.util.HashMap;
