+++ src/test/java/kitchenpos/order/application/OrderTableServiceTest.java
@@ -1,7 +1,7 @@
 package kitchenpos.order.application;
 
-import kitchenpos.order.domain.InMemoryOrderRepository;
-import kitchenpos.order.domain.InMemoryOrderTableRepository;
+import kitchenpos.order.repository.InMemoryOrderRepository;
+import kitchenpos.order.repository.InMemoryOrderTableRepository;
 import kitchenpos.order.domain.OrderRepository;
 import kitchenpos.order.domain.OrderStatus;
 import kitchenpos.order.domain.OrderTable;
