+++ src/test/java/kitchenpos/product/application/ProductServiceTest.java
@@ -1,8 +1,8 @@
 package kitchenpos.product.application;
 
 import kitchenpos.infra.FakePurgomalumClient;
-import kitchenpos.menu.domain.InMemoryMenuRepository;
-import kitchenpos.product.domain.InMemoryProductRepository;
+import kitchenpos.menu.repository.InMemoryMenuRepository;
+import kitchenpos.product.repository.InMemoryProductRepository;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.product.domain.Product;
